# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge OS API
  version: 1.0.0
  description: |
    MCP Platform — project-centric knowledge-work operating system.
    Multiple AI agents (Claude, Codex) safely collaborate on shared workspaces
    backed by Kubernetes PVCs.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /projects:
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
  /projects/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Archive project (soft delete)
      responses:
        '204':
          description: Project archived
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Workspaces ────────────────────────────────────────────────────────────────
  /projects/{id}/workspaces:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a workspace (provisions PVC + directory tree)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List project workspaces
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceResponse'
  /projects/{id}/workspaces/{wid}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: wid
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace details
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Agents ────────────────────────────────────────────────────────────────────
...
  /projects/{id}/timeline:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: listTimeline
      tags: [timeline]
      summary: List timeline events (cursor-paginated)
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: type
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Timeline events page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelinePage'
  /projects/{id}/timeline/{eid}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: eid
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getTimelineEvent
      tags: [timeline]
      summary: Get a specific timeline event
      responses:
        '200':
          description: Timeline event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEventResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{id}/timeline/{eid}/replay:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: eid
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: replayTimelineEvent
      tags: [timeline]
      summary: Replay a reversible timeline event
      responses:
        '200':
          description: Replay initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEventResponse'
...
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [software, content, marketing, migration, research]
        templateHint:
          type: string
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, paused, archived]
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        namespace:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    # ── Workspaces ────────────────────────────────────────────────────────────
    CreateWorkspaceRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [code, content, legacy, target, mapping]
        mode:
          type: string
          enum: [read-write, read-only]
          default: read-write
    WorkspaceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        mode:
          type: string
        path:
          type: string
        pvcName:
          type: string
        fileTree:
          type: array
          items:
            type: string

    # ── Agents ────────────────────────────────────────────────────────────────
...
    TimelineEventResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [project_created, workspace_created, agent_spawned, agent_stopped,
                 changeset_submitted, changeset_approved, changeset_rejected,
                 changeset_applied, changeset_rolled_back, lock_acquired, lock_released,
                 memory_written, memory_deleted, validator_run, terminal_input,
                 pod_started, pod_stopped, cost_recorded, namespace_provisioned,
                 verification_complete]
        payload:
          type: object
        reversible:
          type: boolean
        replayCmd:
          type: string
        agentId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    TimelinePage:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEventResponse'
        nextCursor:
          type: string
        hasMore:
          type: boolean

    # ── Costs ─────────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object